+++ src/entities/post/ui/post-preview/index.tsx
