+++ src/features/post/ui/index.ts
@@ -1,4 +1,6 @@
 export { PostFilter } from "./post-filter"
 export { PostForm } from "./post-form"
 export { PostActionButtons } from "./add-post-button"
-export { PostSearch } from "./post-search"
+export { SearchPost } from "./search-post"
+export { PostFormContainer } from "./post-form-container"
+export { PostTableRow } from "./post-table-row"
