--- src/features/user/ui/index.ts
@@ -1,1 +1,0 @@
-export * from "./user-info"
