+++ src/widgets/common/ui/footer/index.tsx
