+++ src/widgets/common/ui/index.ts
@@ -1,2 +1,2 @@
-export { default as Header } from "./Header"
-export { default as Footer } from "./Footer"
+export { default as Header } from "./header"
+export { default as Footer } from "./footer"
