+++ src/widgets/post/ui/index.ts
@@ -1,5 +1,5 @@
-export { PostDetailDialog } from "./PostDetailDialog"
-export { PostFormDialog } from "./PostFormDialog"
-export { PostsManagerWidget } from "./PostsManagerWidget"
-export { PostsSearchFilter } from "./PostsSearchFilter"
-export { PostsTable } from "./PostsTable"
+export { PostDetailDialog } from "./post-detail-dialog"
+export { PostFormDialog } from "./post-form-dialog"
+export { PostsManagerWidget } from "./posts-manager-widget"
+export { PostsSearchFilter } from "./posts-search-filter"
+export { PostsTable } from "./posts-table"
