+++ src/widgets/post/ui/post-detail-dialog/index.tsx
